refactor file structur for tv series
