+++ domain/vod/src/main/java/details/usecase/tv_series/GetTvSeriesSeasonsUseCase.kt
@@ -8,6 +8,6 @@ class GetTvSeriesSeasonsUseCase(
     private val repo: TvSeriesRepository
 ) {
     suspend fun execute(id: Int, seasonNumber: Int): List<Season> {
-        return repo.getTvSeriesSeasons(id, seasonNumber)
+        return repo.getTvSeriesSeason(id, seasonNumber)
     }
 }
