+++ domain/vod/src/test/java/usecase/details/tv_series/GetEpisodeDetailsUseCaseTest.kt
@@ -1,15 +1,15 @@
-package usecase
+package usecase.details.tv_series
 
 import details.repository.TvSeriesRepository
 import details.usecase.tv_series.GetEpisodeDetailsUseCase
+import entity.Episode
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import kotlin.test.Test
-
 
 class GetEpisodeDetailsUseCaseTest {
 
@@ -24,7 +24,7 @@ class GetEpisodeDetailsUseCaseTest {
     @Test
     fun `execute() should return episode details from TvSeriesRepository`() = runTest {
         // Given
-        val expected = mockk<entity.Episode>()
+        val expected = mockk<Episode>()
         coEvery { repository.getEpisodeDetails(1, 1, 1) } returns expected
 
         // When
@@ -35,16 +35,17 @@ class GetEpisodeDetailsUseCaseTest {
         assert(result == expected)
     }
     @Test
-    fun `execute() should throw exception when repository fails to get episode details`() = runTest {
-        // Given
-        val exception = IllegalStateException("Episode not found")
-        coEvery { repository.getEpisodeDetails(1, 1, 1) } throws exception
+    fun `execute() should throw exception when repository fails to get episode details`() =
+        runTest {
+            // Given
+            val exception = IllegalStateException("Episode not found")
+            coEvery { repository.getEpisodeDetails(1, 1, 1) } throws exception
 
-        // When, Then
-        val result = assertThrows<IllegalStateException> {
-            useCase.execute(1, 1, 1)
+            // When, Then
+            val result = assertThrows<IllegalStateException> {
+                useCase.execute(1, 1, 1)
+            }
+
+            assert(result.message == "Episode not found")
         }
-
-        assert(result.message == "Episode not found")
-    }
 }
