+++ domain/vod/src/test/java/usecase/details/tv_series/GetTvSeriesImagesUseCaseTest.kt
@@ -1,4 +1,4 @@
-package usecase.tv_series
+package usecase.details.tv_series
 
 import details.repository.TvSeriesRepository
 import details.usecase.tv_series.GetTvSeriesImagesUseCase
